pyproject.toml
[project]
name = "anythings-ok"
version = "0.1.0"
description = "Offline multimodal GraphRAG with Neo4j and Ollama"
requires-python = ">=3.13"

dependencies = [
    # Core
    "requests>=2.31",
    "neo4j>=5.0",
    # Text extraction (LlamaIndex handles pdf, docx, pptx, epub, html, etc.)
    "llama-index-core>=0.12",
    "llama-index-readers-file>=0.4",
    # Audio transcription (CPU, no CUDA needed)
    "moonshine-voice>=0.0.49",
    # Graph & community detection
    "networkx>=3.2",
    "igraph>=0.11",
    "leidenalg>=0.10",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "graphdb", "cli"]

[tool.uv]
dev-dependencies = []
